% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_history.R
\name{animate_history}
\alias{animate_history}
\title{Build race-style bar plots of Destiny 2 weapon usage over time}
\usage{
animate_history(
  x,
  activity_type = NULL,
  limit = 30,
  label = "",
  prefix = "d2rstats",
  gif = FALSE,
  mp4 = FALSE
)
}
\arguments{
\item{x}{The D2Rstats weapon-level data set to process}

\item{activity_type}{One activity type integer to filter}

\item{limit}{The total number of top weapons to plot}

\item{label}{The title to put on the plot}

\item{prefix}{Prefix to output filename; only used if gif or av are TRUE}

\item{gif}{Boolean: produce .gif file output?}

\item{av}{Boolean: produce .mov file output?}
}
\value{
If \code{gif} and \code{mp4} are \code{FALSE}, returns animated plot object that can be further manipulated or rendered using \code{animate}
}
\description{
Creates an animated "race" graph showing the growth and change
of total weapon usage through the period of time in the provided weapon-level
activity data.
}
\examples{
levelWeapon |> animate_history(activity_type=82,
  label="Dungeon weapons over time",
  limit=40, mp4=TRUE)
}
