% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2helper.R
\name{filter_mode}
\alias{filter_mode}
\title{Filter to specified activity value}
\usage{
filter_mode(data, mode = NULL)
}
\arguments{
\item{data}{D2Rstats weapon-level data set to operate on}

\item{mode}{Required: Integer representing a Destiny 2 activity type}
}
\value{

}
\description{
This convenience function filters for a specific game type
(mode) in record using dplyr's filter, searching via 'grepl' for the
specified activity mode value in the combined mode value string.
}
\details{
Due to the way that the data set is built from the raw JSON postgame
carnage report files, the \code{modeString} is stored in the D2Rstats data as a
series of activity types concatenated with a pipe character. This function
applies a regular expression to determine if the specified mode value is
found within that concatenated set of values.
}
\examples{
d2data |> filter_mode(mode = 82) # filter to Dungeon activities
}
