% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2helper.R
\name{load_pgcr_data}
\alias{load_pgcr_data}
\title{Load stored PGCR data for a specified membership ID}
\usage{
load_pgcr_data(user_id = NULL)
}
\arguments{
\item{user_id}{Destiny membership ID}
}
\description{
Load stored PGCR data for a specified membership ID
}
