% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2helper.R
\name{raise}
\alias{raise}
\title{Simplify repeated calls to nested elements of PGCR data}
\usage{
raise(.data, unnest, label, displayValue = TRUE)
}
\arguments{
\item{displayValue}{}
}
\value{

}
\description{
Simplify repeated calls to nested elements of PGCR data
}
