% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2helper.R
\name{save_pgcr_data}
\alias{save_pgcr_data}
\title{Save current PGCR data set}
\usage{
save_pgcr_data(id = NULL, data_set = NULL)
}
\description{
Saves data to a directory prepended with current membership ID
}
