% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{weapon_barplot}
\alias{weapon_barplot}
\title{Build barplot of weapon usage}
\usage{
weapon_barplot(
  data = NULL,
  type = NULL,
  tier_weapon = NULL,
  mode = NULL,
  limit = 5,
  ncol = 2,
  text_size = 2,
  title = NULL,
  subtitle = NULL
)
}
\arguments{
\item{subtitle}{}
}
\value{

}
\description{
Build barplot of weapon usage
}
