% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{weapon_treemap}
\alias{weapon_treemap}
\title{Build treemap of weapon usage}
\usage{
weapon_treemap(
  data = NULL,
  type = NULL,
  tier_weapon = NULL,
  mode = NULL,
  limit = 200,
  title = NULL,
  subtitle = NULL
)
}
\arguments{
\item{subtitle}{}
}
\value{

}
\description{
Build treemap of weapon usage
}
