% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2helper.R
\name{raise}
\alias{raise}
\title{Simplify repeated calls to nested elements of PGCR data}
\usage{
raise(.data, unnest, label, displayValue = TRUE)
}
\arguments{
\item{.data}{data frame passed into raise}

\item{unnest}{list element whose \code{basic.displayValue} will be brought up}

\item{label}{name of new field created by raise}

\item{displayValue}{set to FALSE to use raw value rather than the displayValue in the nested json}
}
\description{
Simplify repeated calls to nested elements of PGCR data
}
