% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{weapon_barplot}
\alias{weapon_barplot}
\title{Build barplot of weapon usage}
\usage{
weapon_barplot(
  data = NULL,
  type = NULL,
  tier_weapon = NULL,
  mode = NULL,
  limit = 5,
  ncol = 2,
  text_size = 2,
  title = NULL,
  subtitle = NULL
)
}
\arguments{
\item{data}{data set to use for plot}

\item{type}{weapon type to filter}

\item{tier_weapon}{weapon tier to filter (e.g. "legendary")}

\item{mode}{game mode to filter for}

\item{limit}{maximum number of weapons to include in plot}

\item{ncol}{number of columns to build in a panel of barplots}

\item{text_size}{font size for display}

\item{title}{plot title}

\item{subtitle}{plot subtitle}
}
\value{

}
\description{
Build barplot of weapon usage
}
